chore: rename typemap to sqltype
