+++ internal/exporter/dbml/exporter.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"github.com/artarts36/db-exporter/internal/config"
 	"github.com/artarts36/db-exporter/internal/exporter/exporter"
-	"github.com/artarts36/db-exporter/internal/infrastructure/typemap"
+	"github.com/artarts36/db-exporter/internal/infrastructure/sqltype"
 	"github.com/artarts36/db-exporter/internal/schema"
 	"github.com/artarts36/db-exporter/internal/shared/dbml"
 	"log/slog"
@@ -92,7 +92,7 @@ func (e *Exporter) mapTable(
 	}
 
 	for _, col := range tbl.Columns {
-		typ, err := typemap.TransitSQLType(source, config.DatabaseDriverDBML, col.Type.Value)
+		typ, err := sqltype.TransitSQLType(source, config.DatabaseDriverDBML, col.Type.Value)
 		if err != nil {
 			return nil, nil, fmt.Errorf("failed to map column %q type: %w", col.Name, err)
 		}
